Add separate bundle for thread
Main bundle before: 440kB gzip
Main bundle after: 421.5kB gzip
Newly added Thread bundle: 21.17kB gzip
