+++ src/routes.js
@@ -15,8 +15,6 @@ import ModalRoot from './components/modals/modalRoot';
 import Gallery from './components/gallery';
 import Toasts from './components/toasts';
 import Maintenance from './components/maintenance';
-import LoadingDMs from './views/directMessages/components/loading';
-import LoadingThread from './views/thread/components/loading';
 import { Loading, LoadingScreen } from './components/loading';
 import LoadingDashboard from './views/dashboard/components/dashboardLoading';
 import Composer from './components/composer';
@@ -29,17 +27,12 @@ import Status from './views/status';
 import Login from './views/login';
 
 import DirectMessages from './views/directMessages';
+import Thread from './views/thread';
 
 /* prettier-ignore */
 const Explore = Loadable({
   loader: () => import('./views/explore'/* webpackChunkName: "Explore" */),
   loading: ({ isLoading }) => isLoading && <Loading />,
-});
-
-/* prettier-ignore */
-const Thread = Loadable({
-  loader: () => import('./views/thread'/* webpackChunkName: "Thread" */),
-  loading: ({ isLoading }) => isLoading && <LoadingThread />,
 });
 
 /* prettier-ignore */
